graphToString is specific for SimpleGraph
Also expose the test variables to subclasses
